Use u9 for chunk constant index
When adding constants to a chunk from the parser we expect the index to
be within a u8, so that a single chunk can have no more than 255
constants. For this check to make sense the chunk needs to be able to
return an index larger than u8, so we use u9.